Elevate out classed of IndexBuilder
