+++ WalletWasabi/Blockchain/BlockFilters/History/Operation.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace WalletWasabi.Blockchain.BlockFilters.History
+{
+	public enum Operation
+	{
+		Add,
+		Remove
+	}
+}
